# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Type \"Bearer\" followed by a space and JWT token.",
        "title": "Sales API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/domains": {
            "get": {
                "description": "Get domain list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ドメイン"
                ],
                "summary": "Get domains",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ドメイン名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "閲覧可能か",
                        "name": "can_view",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "日本のサイトか",
                        "name": "is_japan",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "mapsで問い合わせページを開いたか",
                        "name": "is_send",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "owner_id",
                        "name": "owner_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ステータス",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "業種",
                        "name": "industry",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "SSL対応可否",
                        "name": "is_ssl",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Domains"
                            }
                        }
                    }
                }
            }
        },
...
        "/domains/{id}": {
            "get": {
                "description": "Get domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ドメイン"
                ],
                "summary": "Get domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Domain"
                        }
                    }
                }
            },
            "put": {
                "description": "Update domain information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ドメイン"
                ],
                "summary": "Update domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新ドメイン情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateDomainRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Domain"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete domain by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ドメイン"
                ],
                "summary": "Delete domain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/external/api/deploy": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wordpress"
                ],
                "summary": "ワードプレスをデプロイします",
                "parameters": [
                    {
                        "description": "デプロイ情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.DeployRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/fetch": {
            "post": {
                "description": "Fetch domain information from target",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ViewDNS"
                ],
                "summary": "Fetch domains",
                "parameters": [
                    {
                        "description": "Fetch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.PostFetchRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "description": "Get log list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ログ"
                ],
                "summary": "Get logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "処理名",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "カテゴリー",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Log"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create new log",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ログ"
                ],
                "summary": "Create log",
                "parameters": [
                    {
                        "description": "作成ログ情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateLogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Log"
                        }
                    }
                }
            }
        },
        "/targets": {
            "get": {
                "description": "Get target list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ターゲット"
                ],
                "summary": "Get targets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Target"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create new target",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ターゲット"
                ],
                "summary": "Create target",
                "parameters": [
                    {
                        "description": "作成ターゲット情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Target"
                        }
                    }
                }
            }
        },
        "/targets/{id}": {
            "put": {
                "description": "Update target information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ターゲット"
                ],
                "summary": "Update target",
                "parameters": [
                    {
                        "description": "更新ターゲット情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Target"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete target by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ターゲット"
                ],
                "summary": "Delete target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "description": "Get task list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "タスク"
                ],
                "summary": "Get tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create new task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "タスク"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "description": "作成タスク情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Task"
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/execute": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "タスク"
                ],
                "summary": "タスクを実行します",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.CreateTargetRequest": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.DeployRequest": {
            "type": "object",
            "properties": {
                "dst": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Deploy"
                    }
                },
                "src": {
                    "$ref": "#/definitions/domain.Deploy"
                }
            }
        },
        "domain.Log": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.PostFetchRequest": {
            "type": "object",
            "required": [
                "target"
            ],
            "properties": {
                "target": {
                    "type": "string"
                }
            }
        },
...
        "response.Domains": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "domains": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Domain"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}